refactor : 도메인 객체 파일명으로 변경
