+++ src/entities/university/model/univsearch.domain.ts
